# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: PESATRIX Open Education Infrastructure API
  description: Reusable, secure, interoperable digital infrastructure for structured
    educational events and information in low-resource environments. No personal data
    is stored by this system.
  version: 0.1.0
paths:
...
  /api/v1/identities:
    post:
      tags:
      - identities
      summary: Create a student identity reference
      operationId: create_api_v1_identities_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - identities
      summary: List student identity references
      operationId: list_identities_route_api_v1_identities_get
      parameters:
      - name: school_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: School Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_IdentityResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/identities/{identity_id}:
    get:
      tags:
      - identities
      summary: Get a student identity reference
      operationId: get_api_v1_identities__identity_id__get
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - identities
      summary: Update a student identity reference (admin)
      operationId: update_api_v1_identities__identity_id__patch
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identity Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/credentials:
    post:
      tags:
      - credentials
      summary: Issue a credential (QR-capable)
      operationId: issue_api_v1_credentials_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - credentials
      summary: List credentials
      operationId: list_credentials_route_api_v1_credentials_get
      parameters:
      - name: student_identity_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Student Identity Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_CredentialResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/credentials/{credential_id}:
    get:
      tags:
      - credentials
      summary: Get a credential
      operationId: get_api_v1_credentials__credential_id__get
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Credential Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/credentials/{credential_id}/revoke:
    post:
      tags:
      - credentials
      summary: Revoke a credential
      operationId: revoke_api_v1_credentials__credential_id__revoke_post
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Credential Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialRevoke'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/credentials/{credential_id}/verify:
    post:
      tags:
      - credentials
      summary: Verify a credential (online)
      operationId: verify_api_v1_credentials__credential_id__verify_post
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Credential Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialVerifyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/credentials/{credential_id}/qr:
    get:
      tags:
      - credentials
      summary: Render the credential as a QR code (PNG bytes)
      operationId: qr_png_api_v1_credentials__credential_id__qr_get
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Credential Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/events:
    post:
      tags:
      - events
      summary: Record an educational event (idempotent)
      operationId: create_event_api_v1_events_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EducationalEventCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationalEventResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - events
      summary: List educational events
      operationId: list_events_route_api_v1_events_get
      parameters:
      - name: school_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: School Id
      - name: student_identity_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Student Identity Id
      - name: event_kind
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/EventKind'
          - type: 'null'
          title: Event Kind
      - name: since_seq
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Since Seq
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_EducationalEventResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CredentialCreate:
      properties:
        student_identity_id:
          type: string
          format: uuid
          title: Student Identity Id
        issuer_organization_id:
          type: string
          format: uuid
          title: Issuer Organization Id
        credential_type:
          $ref: '#/components/schemas/CredentialType'
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        issued_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Issued At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        payload:
          additionalProperties: true
          type: object
          title: Payload
      type: object
      required:
      - student_identity_id
      - issuer_organization_id
      - credential_type
      - title
      title: CredentialCreate
    CredentialResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        student_identity_id:
          type: string
          format: uuid
          title: Student Identity Id
        credential_type:
          $ref: '#/components/schemas/CredentialType'
        title:
          type: string
          title: Title
        status:
          $ref: '#/components/schemas/CredentialStatus'
        issued_at:
          type: string
          format: date-time
          title: Issued At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        issuer_organization_id:
          type: string
          format: uuid
          title: Issuer Organization Id
        signature:
          type: string
          title: Signature
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        revoke_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Revoke Reason
      type: object
      required:
      - id
      - created_at
      - updated_at
      - student_identity_id
      - credential_type
      - title
      - status
      - issued_at
      - issuer_organization_id
      - signature
      title: CredentialResponse
    CredentialRevoke:
      properties:
        reason:
          type: string
          maxLength: 200
          minLength: 1
          title: Reason
      type: object
      required:
      - reason
      title: CredentialRevoke
    CredentialStatus:
      type: string
      enum:
      - issued
      - revoked
      - expired
      title: CredentialStatus
    CredentialType:
      type: string
      enum:
      - student_id_card
      - enrolment_certificate
      - participation
      - completion
      - other
      title: CredentialType
    CredentialVerifyRequest:
      properties:
        credential_id:
          type: string
          format: uuid
          title: Credential Id
        issuer_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Issuer Code
      type: object
      required:
      - credential_id
      title: CredentialVerifyRequest
    CredentialVerifyResponse:
      properties:
        valid:
          type: boolean
          title: Valid
        status:
          anyOf:
          - $ref: '#/components/schemas/CredentialStatus'
          - type: 'null'
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        student_identity_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Student Identity Id
        credential_type:
          anyOf:
          - $ref: '#/components/schemas/CredentialType'
          - type: 'null'
      type: object
      required:
      - valid
      title: CredentialVerifyResponse
...
    EducationalEventCreate:
      properties:
        school_id:
          type: string
          format: uuid
          title: School Id
        student_identity_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Student Identity Id
        event_kind:
          $ref: '#/components/schemas/EventKind'
        occurred_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Occurred At
        device_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Device Id
        source:
          type: string
          title: Source
          default: api
        idempotency_key:
          anyOf:
          - type: string
            maxLength: 128
            pattern: ^[\w\-]{8,128}$
          - type: 'null'
          title: Idempotency Key
        parent_event_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Event Id
        payload:
          additionalProperties: true
          type: object
          title: Payload
      type: object
      required:
      - school_id
      - event_kind
      title: EducationalEventCreate
    EducationalEventResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        seq:
          type: integer
          title: Seq
        school_id:
          type: string
          format: uuid
          title: School Id
        student_identity_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Student Identity Id
        event_kind:
          $ref: '#/components/schemas/EventKind'
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        recorded_at:
          type: string
          format: date-time
          title: Recorded At
        device_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Device Id
        source:
          type: string
          title: Source
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        parent_event_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Event Id
        version:
          type: integer
          title: Version
        payload:
          additionalProperties: true
          type: object
          title: Payload
      type: object
      required:
      - id
      - created_at
      - updated_at
      - seq
      - school_id
      - event_kind
      - occurred_at
      - recorded_at
      - source
      - version
      - payload
      title: EducationalEventResponse
    EventKind:
      type: string
      enum:
      - check_in
      - check_out
      - lesson_started
      - lesson_completed
      - attendance_override
      - credential_issued
      - credential_verified
      - sync_batch
      title: EventKind
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IdentityCreate:
      properties:
        school_id:
          type: string
          format: uuid
          title: School Id
        code:
          type: string
          maxLength: 64
          minLength: 2
          pattern: ^[a-zA-Z0-9_-]+$
          title: Code
        external_reference:
          anyOf:
          - type: string
          - type: 'null'
          title: External Reference
        status:
          $ref: '#/components/schemas/IdentityStatus'
          default: active
      type: object
      required:
      - school_id
      - code
      title: IdentityCreate
    IdentityResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        school_id:
          type: string
          format: uuid
          title: School Id
        code:
          type: string
          title: Code
        external_reference:
          anyOf:
          - type: string
          - type: 'null'
          title: External Reference
        status:
          $ref: '#/components/schemas/IdentityStatus'
      type: object
      required:
      - id
      - created_at
      - updated_at
      - school_id
      - code
      - status
      title: IdentityResponse
    IdentityStatus:
      type: string
      enum:
      - active
      - suspended
      - archived
      title: IdentityStatus
    IdentityUpdate:
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/IdentityStatus'
          - type: 'null'
        external_reference:
          anyOf:
          - type: string
          - type: 'null'
          title: External Reference
      type: object
      title: IdentityUpdate
...
    PaginatedResponse_CredentialResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CredentialResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[CredentialResponse]
    PaginatedResponse_EducationalEventResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EducationalEventResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[EducationalEventResponse]
    PaginatedResponse_IdentityResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/IdentityResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[IdentityResponse]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError